fix: deduplicate GraphQL queries, sync version, fix CI badge, remove dead hook
- Extract shared GraphQL queries into src/queries.ts used by both
  commands/search.ts and sdk.ts
- Read CLI version from npm_package_version instead of hardcoding 0.1.0
- Fix README CI badge to point to ci.yml instead of npm-publish.yml
- Remove unused empty postAction hook in cli.ts
